refactor:move portfolio details files to portfolios/detail
